+++ src/app/components/date.test.tsx
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 import { DateString } from "./date";
-import { render } from "../utils/react-test";
+import { render } from "../../utils/react-test";
 
 it("renders correctly", () => {
   const tree = render(<DateString dateString="2021-01-01" />);
