+++ src/app/components/footer.test.tsx
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 import { Footer } from "./footer";
-import { render } from "../utils/react-test";
+import { render } from "../../utils/react-test";
 
 it("renders correctly", () => {
   const tree = render(<Footer />);
