+++ src/app/components/header.test.tsx
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 import { Header } from "./header";
-import { render } from "../utils/react-test";
+import { render } from "../../utils/react-test";
 
 it("renders correctly", () => {
   const tree = render(<Header />);
