+++ src/app/components/navigation.test.tsx
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 import { Navigation } from "./navigation";
-import { render } from "../utils/react-test";
+import { render } from "../../utils/react-test";
 
 it("renders correctly", () => {
   const tree = render(
