+++ src/app/components/navigation.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {
   Button,
   Container,
