+++ src/app/components/prev-next-link.test.tsx
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 import { PrevNextLink } from "./prev-next-link";
-import { render } from "../utils/react-test";
+import { render } from "../../utils/react-test";
 
 it("renders both prev and next correctly", () => {
   const tree = render(
