+++ src/app/components/questions.test.tsx
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 import { Questions } from "./questions";
-import { render } from "../utils/react-test";
+import { render } from "../../utils/react-test";
 
 it("renders correctly", () => {
   const tree = render(<Questions />);
