+++ src/app/components/sns-link.test.tsx
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 import { SNSLink } from "./sns-link";
-import { render } from "../utils/react-test";
+import { render } from "../../utils/react-test";
 
 it("renders twitter link correctly", () => {
   const tree = render(<SNSLink type="twitter" url="https://twitter.com/MikurXina" />);
