+++ src/app/components/title.test.tsx
@@ -1,6 +1,6 @@
 import { Subtitle, Title } from "./title";
 import { expect, it } from "vitest";
-import { render } from "../utils/react-test";
+import { render } from "../../utils/react-test";
 
 it("renders title correctly", () => {
   const tree = render(<Title>Epic Title</Title>);
