+++ src/app/fonts.ts
@@ -1,0 +1,6 @@
+import { Roboto } from "next/font/google";
+
+export const roboto = Roboto({
+  weight: "400",
+  subsets: ["latin"],
+});
