--- src/pages/_document.tsx
@@ -1,31 +1,0 @@
-import Document, { Head, Html, Main, NextScript } from "next/document";
-import { ColorModeScript } from "@chakra-ui/react";
-import { theme } from "../lib/theme";
-
-class MyDocument extends Document {
-  render(): JSX.Element {
-    return (
-      <Html lang="ja">
-        <Head>
-          <link
-            rel="stylesheet"
-            href="https://cdn.jsdelivr.net/npm/yakuhanjp@3.3.1/dist/css/yakuhanjp.min.css"
-            media="print"
-          />
-          <link
-            rel="stylesheet"
-            href="https://fonts.googleapis.com/css2?family=Roboto&display=swap"
-            media="print"
-          />
-        </Head>
-        <body>
-          <ColorModeScript initialColorMode={theme.config.initialColorMode} />
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
-
-export default MyDocument;
